+++ src/main/java/com/example/demo/support/domain/CommonSearch.java
@@ -3,57 +3,50 @@ package com.example.demo.support.domain;
 import com.example.demo.dao.Research;
 import com.example.demo.domain.Stock;
 import com.example.demo.domain.StockRepository;
-import org.openqa.selenium.*;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.firefox.FirefoxDriver;
-import org.openqa.selenium.firefox.FirefoxProfile;
-import org.openqa.selenium.phantomjs.PhantomJSDriver;
-import org.openqa.selenium.phantomjs.PhantomJSDriverService;
-import org.openqa.selenium.remote.DesiredCapabilities;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 
-import java.io.File;
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
 
 public abstract class CommonSearch {
     private static final Logger logger =  LoggerFactory.getLogger(CommonSearch.class);
-    private WebDriver driver;
-
-    @Value("${driver.path}")
-    private String path;
-
-    @Value("${phanthomjs.path}")
-    private String phantomPath;
 
     @Autowired
     private StockRepository stockRepository;
 
-//    public void init() {
-//        DesiredCapabilities caps = new DesiredCapabilities();
-//        caps.setCapability(PhantomJSDriverService.PHANTOMJS_EXECUTABLE_PATH_PROPERTY, phantomPath);
-//        driver = new PhantomJSDriver(caps);
-//    }
-
-    public void init() {
-        ChromeOptions options = new ChromeOptions();
-//        options.addArguments("--headless");
-        System.setProperty("webdriver.chrome.driver", path);
-        driver = new ChromeDriver(options);
+    public String [] splitBody(String body) {
+        return body.split("\\s,\\s");
     }
 
-    public void getStart(String url) {
-        driver.get(url);
-        driver.manage().window().setPosition(new Point(0, 0));
-        driver.manage().window().setSize(new Dimension(1360, 430));
-//        JavascriptExecutor js = (JavascriptExecutor) driver;
-//        js.executeScript("return window.stop()");
+    public Stock makingStockUsingJson(String info, JsonParser parser, List<Stock> stocks) {
+        JsonObject object = (JsonObject)parser.parse(info);
+        if (chekcDB(object, stocks)) {
+            logger.info("주식 업데이트");
+            return stockRepository.findByName(getTitle(object)).realDataUpdate(object.get("name").getAsString(), object.get("cost").getAsString(), object.get("updn").getAsString(), object.get("rate").getAsString(), getUrl(object.get("code").getAsString()));
+        }
+        return new Stock(object.get("name").getAsString(), object.get("cost").getAsString(), object.get("updn").getAsString(), object.get("rate").getAsString(), getUrl(object.get("code").getAsString()));
+    }
+
+    public String getUrl(String code) {
+        return "http://finance.daum.net/item/main.daum?code="+code;
+    }
+
+    public String getTitle(JsonObject object) {
+        return object.get("name").getAsString();
+    }
+
+    public boolean chekcDB(JsonObject object, List<Stock> stocks) {
+        logger.info("checkDB에 전달받은 title : {}", getTitle(object));
+        if (stocks.contains(getTitle(object))) {
+            logger.info("db에 {} 존재", getTitle(object));
+            return true;
+        }
+        logger.info("db에 존재안함");
+        return false;
     }
 
     public Stock update(Stock original) throws IOException {
@@ -67,40 +60,4 @@ public abstract class CommonSearch {
         return original;
     }
 
-    public WebDriver getDriver() {
-        return driver;
-    }
-
-    public boolean checkDb(WebElement element, List<Stock> stocks) {
-        if (stocks.contains(getTitle(element)))
-            return true;
-        return false;
-    }
-
-    public List<WebElement> getElements(int partNumber) {
-        return getDriver().findElements(By.cssSelector("#wrap .wBox:nth-child(" + partNumber  +") .nBox div:nth-child(3) > dl"));
-    }
-
-    public Stock making(WebElement element, List<Stock> stocks) {
-        if (checkDb(element, stocks))
-            return stockRepository.findByName(getTitle(element)).realDataUpdate(getTitle(element), getInfo(element).get(1), getInfo(element).get(2), getInfo(element).get(3), getUrl(element));
-        return new Stock(getTitle(element), getInfo(element).get(1), getInfo(element).get(2), getInfo(element).get(3), getUrl(element));
-    }
-
-    public List<WebElement> getChilds(WebElement element) {
-        return element.findElements(By.xpath(".//*"));
-    }
-
-    public String getUrl(WebElement element) {
-        String url = getChilds(element).get(0).getAttribute("onclick");
-        return url.substring(8, url.length() -2);
-    }
-
-    public String getTitle(WebElement element) {
-        return getChilds(element).get(0).getAttribute("title");
-    }
-
-    public List<String> getInfo(WebElement element) {
-        return Arrays.asList(element.getText().split("\n"));
-    }
 }
